// src/coins/ethereumLike/types.ts
export interface AbiItem {
  anonymous?: boolean
  constant?: boolean
  inputs?: AbiInput[]
  name?: string
  outputs?: AbiOutput[]
  payable?: boolean
  stateMutability?: StateMutabilityType
  type: AbiType
  gas?: number
}

export interface AbiInput {
  name: string
  type: string
  indexed?: boolean
  components?: AbiInput[]
  internalType?: string
}

export interface AbiOutput {
  name: string
  type: string
  components?: AbiOutput[]
  internalType?: string
}

export type AbiType = 'function' | 'constructor' | 'event' | 'fallback'
export type StateMutabilityType = 'pure' | 'view' | 'nonpayable' | 'payable'

export type TChain = {
  chain: string
  explorerLink: string
  tokenExplorerLink: string
  txLink: string
}